bot/requirements.txt
amqp==5.0.9
anyjson==0.3.3
APScheduler==3.6.3
apturl==0.5.2
asgiref==3.5.2
asn1crypto==1.4.0
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
Brlapi==0.8.3
cachetools==4.2.2
celery==5.2.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.2
django-birthday==0.1.4
django-cors-headers==3.13.0
django-countries==7.4.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
duplicity==0.8.21
ephem==4.1.3
fasteners==0.14.1
filelock==3.8.1
Flask==2.2.2
future==0.18.2
gunicorn==20.1.0
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
kazam==1.4.5
keyring==23.5.0
kombu==5.2.3
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
mechanize==0.4.7
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
pandas==1.5.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.28
protobuf==3.12.4
protonvpn-cli==2.2.11
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySnooper==1.1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-memcached==1.58
python-telegram-bot==13.14
pythondialog==3.5.3
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.26.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.3
sshuttle==1.0.5
systemd-python==234
tornado==6.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.17.0
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
xdg==5
xkit==0.0.0
zipp==1.0.0
